renaming components/dataset_generator.py to components/generator.py
