+++ src/datumaro/components/environment.py
@@ -79,7 +79,7 @@ class Environment:
             SubsetBase,
             Transform,
         )
-        from datumaro.components.dataset_generator import DatasetGenerator
+        from datumaro.components.generator import DatasetGenerator
         from datumaro.components.exporter import Exporter
         from datumaro.components.launcher import Launcher
         from datumaro.components.validator import Validator
