+++ src/datumaro/components/generator.py
